# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  displayName: string;
}

export interface AdminLoginResponse {
  user: AdminUser;
}

export interface ErrorMessage {
  message: string;
}