# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      description: Authenticate an admin user and create a session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      description: Destroy the current admin session
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: adminMe
      tags: [auth]
      summary: Current admin
      description: Returns the currently authenticated admin user
      responses:
        "200":
          description: Authenticated admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
      required:
        - id
        - username
        - displayName
    AdminLoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - user
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message