# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Authenticate an admin user and create a session cookie
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  "username": zod.string(),
  "password": zod.string()
})

export const AdminLoginResponse = zod.object({
  "user": zod.object({
  "id": zod.number(),
  "username": zod.string(),
  "displayName": zod.string()
})
})


/**
 * Returns the currently authenticated admin user
 * @summary Current admin
 */
export const AdminMeResponse = zod.object({
  "id": zod.number(),
  "username": zod.string(),
  "displayName": zod.string()
})